+++ packages/terra-button/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated default margin value for terra-button.
+
 ## 3.74.0 - (February 15, 2024)
 
 * Changed