+++ packages/terra-core-docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 ## 1.59.0 - (February 23, 2024)
 
 * Added
+  * Added a test example for `terra-button` to validate focus inside table.
   * Added test to `terra-section-header` for the sticky section header title.
 
 ## 1.58.0 - (February 20, 2024)